requirements.txt
adrf==0.1.1
asgiref==3.7.2
certifi==2023.5.7
click==8.1.3
cloudinary==1.33.0
Django==4.2.2
django-autoslug==1.9.9
django-cors-headers==4.1.0
django-debug-toolbar==4.1.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==18.11.1
h11==0.14.0
inflection==0.5.1
iniconfig==2.0.0
packaging==23.1
Pillow==9.5.0
pluggy==1.2.0
psycopg-binary==3.1.9
PyJWT==2.7.0
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.22.0
whitenoise==6.5.0
